pyproject.toml
[tool.poetry]
name = "fastapijwt"
version = "0.1.0"
description = ""
authors = ["Artem Muzychuk <muzychukartem.music@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
SQLAlchemy = {extras = ["asyncio"], version = "^2.0.31"}
aiosqlite = "^0.20.0"
uvicorn = "^0.29.0"
pydantic = "^2.7.1"
pydantic-settings = "^2.2.1"
greenlet = "^3.0.3"
alembic = "^1.13.1"
pyjwt = {extras = ["crypto"], version = "^2.8.0"}
bcrypt = "^4.1.3"
python-multipart = "^0.0.9"
psycopg2 = "^2.9.9"
asyncpg = "^0.29.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"